requirements.txt
requests
pillow
flask
flask_socketio
opencv-python
datetime